feat: allow reuse of workflows by other organizations
Problem statement
=================
When using workflows such as:
  * contributors
  * bump-release
  * auto-merge

the retrieval of secrets for commit or tag PGP-signature and
token switch with a github app is currently specific to go-openapi.

Proposed solution
=================
The names of the secrets (not the secrets themselves) can be injected
via optional input parameters into these shared workflows.

To avoid excessive secret exposure in workflows, usage of the injected secrets
is handed over to a dedicated action, that configures GPG (for signing
secrets) or switches token (for github app token exchange).

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>