fix: add a delay before the contributors workflow sets auto-merge
Problem: when auto-merge is set, the PR merges as soon as the required check
status is okay, meaning that the next job that uploads test coverage
always fails (branch is deleted). This is not blocking but kind of untidy.

Solution: introduced a loop that waits for all PR jobs to complete
before auto-merge is enabled.

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>