feat(release): added support for remote shared config to be used to configure git-cliff
This change affects the release.yml workflow by bringing in new options.

Existing behavior is not modified, thanks to default input values
(.cliff.toml).

New feature:
If no local .cliff.toml config is found (or any specific local config
file provided as input), a remote config is used.

The default is to use as a shared config the .cliff.toml of this repo on
master.

* fixes #36

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>